Extract power function into dedicated module
Move power and powerBySquaring from operations.ts to
src/functions/arithmetic/power.ts for better separation of concerns.
